Move exit-drain tests to their own file
worker_threads.test.ts has pre-existing debug-mode timeouts in unrelated
subprocess-spawn tests; a dedicated file gives the new coverage a clean
pass/fail signal.
